[C-4322] Add mobile server collection page
