+++ packages/web/src/ssr/SsrContext.tsx
@@ -1,7 +1,7 @@
 import { createContext, useContext } from 'react'
 
-import { SsrPageProps } from '@audius/common/models'
-import { Nullable } from '@audius/common/utils'
+import type { SsrPageProps } from '@audius/common/models'
+import type { Nullable } from '@audius/common/utils'
 import { History } from 'history'
 
 import { isMobile as isMobileClient } from 'utils/clientUtil'
@@ -24,12 +24,12 @@ export type SsrContextType = {
    * The page props for the current request. This is available on both the server and the client.
    * For example, this can contain track data for rendering the track page
    */
-  pageProps: SsrPageProps
+  pageProps?: SsrPageProps
   /**
    * The history object for the current request. This is only available on the server.
    * Use useHistoryContext to access the history object on the client.
    */
-  history: Nullable<History>
+  history?: Nullable<History>
 }
 
 export const useSsrContext = () => {
